Se creo la vista de las asistencias, se cambio el logo, se agrego la ruta de perfil en el panel de admin, y se cambio las rutas de edit y create por editar y crear

+++ app/Http/Controllers/Admin/ProfesoreController.php
@@ -84,7 +84,10 @@ class ProfesoreController extends Controller
             $uri = Storage::put('pdfs',$request->file('cv'));
 
             if($profesore->cv){
-                Storage::delete($profesore->cv);
+
+                if(Storage::exists($profesore->cv)){
+                    Storage::delete($profesore->cv);
+                }
 
                 $profesore->update([
                     'cv' => $uri,
